+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet.
+### Changed
+- Swapped to use goccy/go-json for improved performance. Thanks @imix
 
 ## [v3.2.1] - 2026-01-05