+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Swapped to use goccy/go-json for improved performance. Thanks @imix
+- Updated model `IsScalar` internals to improve efficiency. Thanks @imix
 
 ## [v3.2.1] - 2026-01-05