+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - XML parser now preserves comments and processing instructions during round-trip (#175).
 
+### Changed
+- Swapped to use goccy/go-json for improved performance. Thanks @imix
+- Updated model `IsScalar` internals to improve efficiency. Thanks @imix
+
 ## [v3.2.1] - 2026-01-05
 
 ### Fixed